Pet Classification Model Using CNN
The objective of the Pet Classification Model Using CNN (Convolutional Neural Network) project is to develop a deep learning model that accurately classifies different breeds of pets(cats, Dogs) from images. By analyzing the visual features and patterns in the images, the project aims to create a model that can automatically recognize and classify  cats, Dogs with high accuracy.